requirements.txt
asgiref==3.8.1
attrs==23.2.0
backports.zoneinfo==0.2.1
Django==4.2.13
djangorestframework==3.15.1
drf-spectacular==0.27.2
flake8==7.0.0
importlib-resources==6.4.0
inflection==0.5.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
mccabe==0.7.0
pkgutil-resolve-name==1.3.10
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyflakes==3.2.0
PyYAML==6.0.1
referencing==0.35.1
rpds-py==0.18.1
sqlparse==0.5.0
typing-extensions==4.11.0
uritemplate==4.1.1
zipp==3.18.1
